+++ Lib/fontbakery/checks/varfont/bold_wght_coord.py
@@ -2,17 +2,16 @@ from fontbakery.prelude import check, Message, FAIL, PASS, SKIP
 
 
 @check(
-    id="googlefonts/varfont/bold_wght_coord",
+    id="varfont/bold_wght_coord",
     rationale="""
         The OpenType spec's registered
         design-variation tag 'wght' available at
         https://docs.microsoft.com/en-gb/typography/opentype/spec/dvaraxistag_wght
         does not specify a required value for the 'Bold' instance of a variable font.
 
-        But Dave Crossland suggested that we should enforce
-        a required value of 700 in this case (NOTE: a distinction
-        is made between "no bold instance present" vs "bold instance is present
-        but its wght coordinate is not == 700").
+        But Dave Crossland suggested that a required value of 700 should be enforced 
+        in this case (NOTE: a distinction is made between "no bold instance present"
+        vs "bold instance is present but its wght coordinate is not == 700").
     """,
     conditions=["is_variable_font", "has_wght_axis"],
     proposal="https://github.com/fonttools/fontbakery/issues/1707",